requirements.txt
pdf2docx==0.4.2
docx2pdf==0.1.8
sounddevice==0.4.1
scipy==1.7.1
pillow==8.3.1
rembg==1.1.4
easyocr==1.3.1
gtts==2.2.3
PyPDF2==1.26.0
pyttsx3==2.90
